// src/types/sdk_interfaces.ts
export enum ErrorCode {
  /**
   * @description The init options object is invalid.
   */
  InvalidInitOptions = "invalid_initialization_options",
  /**
   * @description The sdk is not initialized.
   */
  NotInitialized = "not_initialized",
  /**
   * @description There is no active Journey.
   */
  NoActiveJourney = "no_active_journey",
  /**
   * @description Unable to receive response from the server.
   */
  NetworkError = "network_error",
  /**
   * @description The client response to the Journey is not valid.
   */
  ClientResponseNotValid = "client_response_not_valid",
  /**
   * @description The server returned an unexpected error.
   */
  ServerError = "server_error",
  /**
   * @description The provided state is not valid for SDK state recovery.
   */
  InvalidState = "invalid_state",

  /**
   * @description The provided credentials are invalid.
   */
  InvalidCredentials = "invalid_credentials",

    /**
     * @description The provided OTP passcode is expired.
     */
  ExpiredOTPPasscode = "expired_otp_passcode",

  /**
   * @description The provided validation passcode is expired.
   */
  ExpiredValidationPasscode = "expired_validation_passcode",


  /**
   * @description Max resend attempts reached
   */
  MaxResendReached = "expired_otp_passcode",
}

/**
 * @interface
 * @description Common interface for Promise rejections. Developers should handle according to the @errorCode
 */
export interface IdoSdkError {
  /**
   * @description The error code.
   */
  readonly errorCode: ErrorCode;
  /**
   * @description The error description.
   */
  readonly description: string;
  /**
   * @description The error additional data. Optional.
   */
  readonly data?: any;
}

/**
 * @enum
 * @description The enum for the client response option types.
 */
export enum ClientResponseOptionType {
  /**
   * @description Client response option type for client input. This is the standard response option for any step.
   */
  ClientInput = "client_input",
    /**
   * @description Client response option type for a cancelation branch in the Journey. Use this for canceling the current step.
   */
  Cancel = "cancel",
  /**
   * @description Client response option type for a failure branch in the Journey. Use this for reporting client side failure for the current step.
   */
  Fail = "failure",
  /**
   * @description Client response option type for custom branch in the Journey, used for custom branching.
   */
  Custom = "custom",

  /**
   * @description Client response option type for a resend of the OTP. Use this for restarting the current step (sms / email otp authentication).
   */
  Resend = "resend",
}

/**
 * @interface
 * @description The interface for client response option object. Use this object to submit client input to the Journey
 * step to process, cancel the current step or choose a custom branch.
 */
export interface ClientResponseOption {
  /**
   * @description The type of the client response option.
   */
  readonly type: ClientResponseOptionType;
  /**
   * @description The id of the client response option.
   * Journey step unique id is provided for the {@link ClientResponseOptionType.Custom} response option type.
   * {@link ClientResponseOptionType.ClientInput} and {@link ClientResponseOptionType.Cancel} have standard Ids _ClientInput_ and _Cancel_, respectively.
   */
  readonly id: string;
  /**
   * @description The label of the client response option.
   */
  readonly label: string;
  /**
   * @description Optional schema object that can be used for UI rendering.
   */
  schema?: Record<string, any>;
}

/**
 * @deprecated
 * @enum
 * @description Deprecated enum. Use {@link IdoJourneyActionType} instead to detect completion, rejection, or a step that requires client input.
 */
export enum IdoServiceResponseType {
  /**
   * @description The Journey ended successfully.
   */
  JourneySuccess = "journey_success",
  /**
   * @description The Journey reached a step that requires client input.
   */
  ClientInputRequired = "client_input_required",
  /**
   * @description The current Journey step updated the client data or provided an error message.
   */
  ClientInputUpdateRequired = "client_input_update_required",
  /**
   * @description The Journey ended with explicit rejection.
   */
  JourneyRejection = "journey_rejection",
}
// ...
export interface IdoServiceResponse {
  /**
   * @deprecated
   * @description Deprecated attribute. Use {@link IdoJourneyActionType} instead.
   */
  readonly type: IdoServiceResponseType;
  /**
   * @description Optional data object returned from the server for any of the journey steps.
   */
  readonly data?: any;
  /**
   * @description Additional error data returned from the server for any of the journey steps.
   */
  readonly errorData?: IdoSdkError;
  /**
   * @description Contains the Journey step ID, allowing the client side to choose the correct handler and UI.
   * This will be either a form ID for the "Get Information from Client" and "Login Form" journey steps,
   * or one of {@link IdoJourneyActionType} for other actions.
   */
  readonly journeyStepId?: IdoJourneyActionType | string;
  /**
   * @description The Journey client response options if the response type is {@link IdoServiceResponseType.ClientInputRequired}
   * or {@link IdoServiceResponseType.ClientInputUpdateRequired}.
   */
  readonly clientResponseOptions?: Record<
    ClientResponseOptionType | string,
    ClientResponseOption
  >;

  /**
   * @description A proof of journey completion is provided upon successful completion of the journey,
   * indicated by the {@link IdoJourneyActionType.Success} step ID.
   */
  token?: string;

  /**
   * @description If a browser-redirection is required (for example at the end of an SSO journey) - the server will provide the redirect URL here.
   * The client should redirect the browser to this URL, i.e. by issuing a `window.location.href = response.redirectUrl;`
   */
  redirectUrl?: string;
}